board/lemaker/bananapro: fix wlan module
Restore a patch which disables 00B IRQ for the built-in wlan module which
allows working wlan module again.  It turned out that it shouldn't have been
deleted because the bug still exists in the mainline linux kernel, so keep
this patch the same as armbian[1] to have wifi connection working correctly.

[1] https://github.com/armbian/build/blob/master/patch/kernel/sunxi-dev/ARM-dts-sun7i-Disable-OOB-IRQ-for-brcm-wifi-on-Cubietruck-and-Banana-Pro.patch

Signed-off-by: Bartosz Bilas <b.bilas@grinn-global.com>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>