package/snort: needs zlib
In commit 69ee49a2c4872cae81c84d0b014be2d97bc31f19 ("package/libpcap:
remove unnecessary dependency on zlib"), the depedency of libpcap on
zlib was removed as it was not needed. However, it was thanks to this
dependency that the snort package satisfied its dependency on zlib,
which is now missing. This commit fixes that by adding a dependency of
snort on zlib.

Fixes:

  checking zlib.h usability... no
  checking zlib.h presence... no
  checking for zlib.h... no

     ERROR!  zlib header not found, go get it from
     http://www.zlib.net

Signed-off-by: Peter Seiderer <ps.report@gmx.net>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>