package/busybox: change how its SELinux support is enabled
Following commit c287d66b6ad779505c87f79a472f2d211aa6f255
("package/refpolicy: remove dependency on policycoreutils"), we have a
build failure on some configurations:

Makefile:571: *** libselinux is in the dependency chain of busybox that has added it to its _DEPENDENCIES variable without selecting it or depending on it from Config.in.  Stop.

This is because refpolicy selects the busybox SELinux support when
Busybox is enabled, which it turns selects libselinux, but we no
longer pay attention to the libselinux dependencies while doing this.

Since it's quite weird to have refpolicy mess with Busybox SELinux
support, this commit changes the logic to have Busybox automatically
enable its SELinux support as soon as SELinux support is enabled,
while still allowing it to be disabled.

Fixes:

  http://autobuild.buildroot.net/results/5d8fda7c488a03c14942d87467d501acd633d24a/

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>