package/norm: switch to Github hosted archive
The original download server (downloads.pf.itd.nrl.navy.mil)
is quite unreliable, which lead to the download being very slow
or even failing. Since the project is now hosted on Github, we
switch to that site, which makes the download tractable.

Update the hash, and use the opportunity to realign fields to the new
spacing convention.

Signed-off-by: Titouan Christophe <titouan.christophe@railnova.eu>
[yann.morin.1998@free.fr: spaces in hash file]
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>