package/python-iptables: use installed iptables by default
python-iptables depends on ctypes.util.find_library() which does not
work due to the absence of gcc and friends on target. The location of
the xtables library and the iptables modules can be configured through
environment variables. Within the scope of buildroot we can determine
what these should be at build time and replace the calls to os.getenv()
with the correct value.

Fixes: https://bugs.busybox.net/show_bug.cgi?id=12271

Signed-off-by: Frank Vanbever <frank.vanbever@essensium.com>
Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>